container/requirements.txt
pytest
pytest-cov
aws-cdk-lib<3.0.0,>=2.44.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-batch-alpha==2.35.0a0
pre-commit
